restartround no longer instantly ends the game. Instead it fills the usecase most people would expect where it starts the round-end timer and shows the credits.
