--- Content.Server/GameTicking/Commands/NewRoundCommand.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using Content.Server.Administration;
-using Content.Shared.Administration;
-using Robust.Shared.Console;
-using Robust.Shared.GameObjects;
-using Robust.Shared.IoC;
-
-namespace Content.Server.GameTicking.Commands
-{
-    [AdminCommand(AdminFlags.Server)]
-    public class NewRoundCommand : IConsoleCommand
-    {
-        public string Command => "restartround";
-        public string Description => "Moves the server from PostRound to a new PreRoundLobby.";
-        public string Help => String.Empty;
-
-        public void Execute(IConsoleShell shell, string argStr, string[] args)
-        {
-            var ticker = EntitySystem.Get<GameTicker>();
-            ticker.RestartRound();
-        }
-    }
-}
